src/requirements.txt
python-telegram-bot
bidict
jsonpickle
requests
lxml